ast: Let astNewCONV() do all CONST overflow checking
This simplifies quite some code, because several astCheckConst() calls
and even the whole hTruncateInt() can be removed (assuming astCheckConst()
will already do the same checks).

As can be seen with tests/warnings/, many double-warnings are eliminated
by this, but unfortunately due to the change in ast-node-arg.bas the
overflow warning for arguments will no longer be shown via hParamWarning()
currently.