pp: Disallow identifiers/keywords as-is in PP expressions
in favour of true string literals, and switching to cExpression() (in the
next commit) which then will allow using "const foo = 123" constants and
other kinds of constant evaluation.

This wouldn't be possible while identifiers and keywords are silently
converted to string literals in PP expressions, so it seems best to remove
that feature. It's still possible to use a <#define stringify(s) #s> macro
to do it manually if needed.
(PP expressions supporting the same keywords that normal expressions do
seems to me like it's more useful than the special identifiers/keywords
as string literal support)