pp: Use cExpression() to parse PP expressions
This allows all usual FB expressions also in the context of #if, i.e. #if
can use any constant expression instead of being restricted to a subset.
For example, this adds support for AndAlso, OrElse, Iif(), Len(),
Ubound/Lbound(), constants, etc. (basically all operations that the old
PP expression parser didn't support, but cExpression() does, and also any
expressions that are evaluated at compile-time)

Instead of calculating the #if expression value manually, it now relies
on the AST's constant folding like everything else. This removes lots of
duplicate code, thus eliminating potential sources of bugs, and making the
64bit port easier.

In fact, this change fixes some quirks with #if's expression evaluation.

For example, this was evaluated to "false" before this patch, even though
it should be "true":

	#if (1/2) > .49
		#print "(1/2) > .49  ->  true"
	#else
		#print "(1/2) > .49  ->  false"
	#endif

same for this kind of expression involving unsigned arguments:

	#if &hFFFFFFFFu > 0u
		#print "&hFFFFFFFFu > 0u  ->  true"
	#else
		#print "&hFFFFFFFFu > 0u  ->  false"
	#endif

since now the AST handles the calculations, at least the behaviour will
be consistent with the rest of FB now.