symb: Also update a proc's real dtype when back-patching forward refs
When replacing the FWDREF dtype on a symbol with the real dtype (when the
forward reference is "implemented"), it's important to update any data that
depends on the symbol's type, such as the FBSYMBOL.lgt length field,
and for procedures, the FBSYMBOL.proc.realdtype/realsubtype fields.

The real dtype fields on procedure symbols were not updated yet, this patch
fixes that. Normally function results cannot be forward references, but
this has changed now that BYREF results are supported.