Re-add FOR/SELECT/WITH implicit var unscoping in -lang qb/fblite
This is a better fix for the problem that bf5363bc attempted to fix,
and fixes a missing "branch crossing" regression in -lang fb.

In -lang qb/fblite, the implicit vars used by FOR loops should be unscoped,
like explicit vars (those DIM'ed by the user) and temp vars (1 statement),
as FOR loops span across multiple statements and the implicit vars need to
be accessed repeatedly. Unscoping them means they'll use unique (on the
procedure-level) stack slots, instead of possibly being overwritten with
vars from parallel scopes. This allows GOTO/GOSUB in to or out of the FOR
block safely.

Same for WITH and SELECT, although, technically, for SELECT on integers,
the implicit integer var won't be accessed anymore once a CASE block is
reached, so it doesn't matter if the memory slot is re-used during a
GOSUB/RETURN for example, but for SELECT on strings it matters, because
the implicit string var must be freed (fb_StrDelete()) at scope breaks.

When unscoping any implicit vars, corresponding initialization code should
be added to the top of the procedure too (astAddUnscoped()), as done for
explicit vars, that's most important for unscoped implicit string vars,
due to the fb_StrDelete() calls that will be added to any exits from the
procedure.

In -lang fb, it seems to me like implicit vars should not be unscoped.
This again matches the behaviour of explicit and temp vars of the -lang
mode. Afterall, -lang fb has scopes, and GOTO works like EXIT DO & co,
i.e. it inserts a true scope break that destroys locals. Given that, it's
never been ok to GOTO out of a scope and back in. While it is technically
no problem with integer vars, it's madness with strings or or UDTs that
need to be constructed and destructed. At least, GOTOing into a scope
doesn't ensure to call any constructors for "skipped" var declarations;
instead we have "branch crossing" errors that disallow such cases, and the
"branch crossing" warnings for integer vars & co. Besides that, GOSUB isn't
allowed in -lang fb.