Aspell header: Assume the lib is called libaspell.[dll.]a on all systems
When compiling Aspell for MinGW, the produced DLL is called
libaspell-15.dll, the import lib is called libaspell.dll.a and the static
lib is called libaspell.a. Thus we shouldn't #inclib "aspell-15".

Also, the aspell-15.def file can be removed, as the Aspell build process
produces an import lib already.