FB-win32: Build Aspell DLL manually
When building Aspell natively with MinGW (on Win32), I couldn't get it to
produce a dll, despite the use of --enable-shared. (similar to libjasper)
It's probably an issue with the libtool in the MinGW/MSYS setup, or an
issue with the autotools files in the Aspell package itself.

When cross-compiling Aspell from Linux for MinGW, the DLL is produced just
fine though.