Update jpeglib headers and fix the example
The new jpeglib.bi header supports jpeglib 6.2, 7.0, 8.4 and 9.0.
The version can be selected through the __JPEGLIB_VER__ #define,
and it defaults to 9.0 (the version that will be included with
FB release packages).

Supporting multiple versions is necessary because they are ABI
incompatible and e.g. Ubuntu 13.04 apparently has versions 6 and 8,
but not the latest version 9. But it will be different for other
distros, not to mention DOS/Win32...