Unix rtlib: Use _Exit() instead of exit() from forked child process
This fixes a hang due to pthread_join() in src/rtlib/unix/hinit.c waiting
on the pthread handle for the bg_thread() from the parent process. fork()
doesn't duplicate other threads, so the pthread handle was invalid, and
the child process shouldn't wait on it.

I observed this hang only with musl libc, not with glibc, but that may
just be coincidence (or glibc having additional error checking).

Overall it seems to be important to use _Exit() instead of exit() for other
reasons too:
	http://stackoverflow.com/questions/2329640/how-to-exit-a-child-process-exit-vs-exit
	http://stackoverflow.com/questions/5422831/what-is-the-difference-between-using-exit-exit-in-a-conventional-linux-fo
	http://www.unixguide.net/unix/programming/1.1.3.shtml