makefile: Produce fbc.exe, not fbc-new.exe
Since libfb.a etc. are also immediately put into lib/<target>/, without
being renamed, it's better to do the same with fbc.exe too. It can even
be dangerous to always compile fbc-new.exe only, and leave the old fbc.exe
there, while the libs are new ones too, and potentially incompatible with
the old fbc.exe.

Since fbc.exe is used to link the new fbc.exe, it cannot overwrite itself
while it's running (at least with the standalone setup, when not using an
external fbc.exe), and a temporary fbc-new.exe must still be used, but
after the link it can immediately be renamed to fbc.exe (and the makefile
will do that automatically now).