gfxlib2: Check glXCreateContext() result with != 0 instead of > 0
According to man pages found on the web, glXCreateContext() returns NULL
on failure. Also, the GLXContext it returns is a pointer type. Thus != 0
should be better than > 0, and casting to intptr_t and then comparing > 0
especially seems weird since that could cause non-NULL values to be
rejected because they appear to be < 0 (address values > 2GiB wrapping
around/overflowing).