parser: Allow TYPEOF() on SUB|FUNCTION(...)
This extends the TYPEOF() type vs. expression ambiguation to recognize
SUB|FUNCTION followed by '(' as type, just as was already done for TYPEOF
followed by '('.

There should be no breakage due to this change, since it's not allowed to
declare any "toplevel" functions/macros/arrays named SUB or FUNCTION.
Only fields in non-class UDTs can be named SUB|FUNCTION|TYPEOF, but those
would have a 'myudt.' or at least an '.' preceding them, which would be
enough to prevent cTypeOrExpression() from treating it as type.

This could be useful to declare pointers to function pointers directly,
without the help of an intermediate typedef:

	dim pp as typeof( function() as integer ) ptr

is the same as

	type myFuncPtr as function() as integer
	dim pp as myFuncPtr ptr

in any case, it cannot be written as

	dim pp as function() as integer ptr

because then the PTR would be treated as part of the function result type,
i.e. it would be a function pointer returning an integer pointer, not a
pointer to a function pointer returning an integer.