ir: Simplify procAllocArg/Local interface
Length parameter and offset result are only used by ir-tac/emit_x86,
but not the C or LLVM backends. Since the length is calculated from
the symbol, and the offset is stored into the symbol, and the symbol
is passed anyways already, emit_x86 can use the symbol directly,
as it already does with the proc symbol anyways.