symb-proc: Use correct length for struct function result param var
For functions returning structures, an implicit BYREF AS ThatStruct
parameter is added (although only the VAR representing the parameter,
no PARAM symbol itself). It should use the dtype's lengths, not the
pointer size just because it's BYREF, to follow the behaviour for other
parameters. The PARAM symbol uses the "real" length as the parameter
appears on stack, while the corresponding VAR symbol uses the dtype's
length. See also symbAddVarForParam().