Unix rtlib: Don't call non-reentrant functions from SIGWINCH handler
in other words: we probably shouldn't try to call fb_hTermQuery() from
a signal handler because it might have interrupted other fb_hTermQuery()
or fb_hTermOut() calls and the order could get mixed up.

Also, functions such as ioctl(), calloc(), fflush(), fprintf(), fputs(),
tputs(), fscanf() may not be reentrant (i.e. if a signal interrupts
execution of one of them, our signal handler can not safely call it
anymore).

This patch changes the SIGWINCH signal handler to just set a flag,
which is checked by the new fb_hRecheckConsoleSize() function which then
queries the terminal size just like the old SIGWINCH handler used to do.
This way the re-entrancy issues are avoided at the cost of additional
fb_hRecheckConsoleSize() calls that should be done whereever the
__fb_con.{w,h,cur_x,cur_y,char_buffer,attr_buffer} fields need to be
accessed, to ensure they are up-to-date.