makefile: Support multilib too
For example, on a 32bit system, use
	make rtlib gfxlib2 FBMULTILIB=64
to compile 64bit versions of the rtlib/gfxlib into lib/freebasic64/,
using gcc -m64, for usage with
	bin/fbc -arch 64
for cross-compiling from 32bit to 64bit.

This is opposed to using a full cross-compiler, such as:
	make rtlib gfxlib2 TARGET=x86_64-pc-linux-gnu
because multilib toolchains are easily available on current GNU/Linux
distros, while full cross-compilers aren't.