tests: Add test case for function pointer expression subtraction
http://www.freebasic.net/forum/viewtopic.php?f=3&t=21356

Following up with the previous commit: Creating the proper procptr subtype
symbol will now also let it be re-used for function pointers with the same
type in the same scope, even if they were based on a different procedure.

That's how it should be, because type checks such as for pointer arithmetic
in astNewBOP(), compare the subtype by address. Where previously an
expression like @sub1 - @sub2 was refused because the subtypes were
different (pointing to the PROC symbols for sub1 and sub2 respectively),
they will now be mapped to the same procptr subtype symbol, and astNewBOP()
will accept the expression.