Disallow procedure pointer arithmetic
because it doesn't make sense. Previously, procedures were treated as
being 4 bytes in size in the context of pointer arithmetic
(symbCalcDerefLen() and its callers), but that's useless.
Doing +1 (or rather, +4) on a procedure pointer doesn't take it to the
next procedure. Same goes for subtraction of function pointers.
Furthermore, sizeof() doesn't work on functions...