compiler: Revert targets to represent OS only, re-add -arch 32/64
Even though some CPUs may support both 32 and 64 bits, it seems best to
not follow gcc on this. It'd be pretty confusing to have -arch x86_64
and still compile for 32bit. It is much more useful to rely on -arch more
heavily.

If we ever want to use -gen gcc to compile with gcc -march=k8 with either
32bit or 64bit, then we can aswell add -arch k8-32 and -arch k8-64 options.

-target win32|dos|... will serve only as target OS selection; although
of course the normal fbc build still recognizes triplets such as
-target i686-pc-mingw32 which can set both OS and arch.

Another good reason to keep using targets for OS selection mostly,
and keep cputype separate, is that with FB being ported to ARM, there will
be lots of new -arch arm* options. Having separate targets for all of them
would be insane.