Fix ICE in symbLookupAt() on Enum declared in Extern block
symbAddEnum() currently won't allocate the enum's hashtb when
not using FB_MANGLING_BASIC, so symbLookupAt(), or rather the
hashLookup(), would just crash due to divison by zero because
the hashtb has zero size.

The crash on
	dim shared enum1 as Enum1
happened because the compiler treated enum1 as namespace prefix,
reported a missing '.' error, and then tried to lookup the AS
in the Enum1 namespace.