Overload resolution: Refuse z/wstring <-> integer, unless it's a DEREF
z/wstring belong to FB_DATACLASS_INTEGER, and indeed they can be treated
as integers (hence char/wchar), but only if it's a DEREF expression.
In other cases (z/wstring variables or literals) the compiler refuses
treating them as integers, and only allows them to be treated as strings.

This patch adjusts hCalcTypesDiff() to follow those rules more closely,
or so I hope.

- Disallow integer (non-z/wstring) arg to z/wstring param
- Only allow z/wstring arg to integer param if it's a DEREF
- Allow string arg to z/wstring param