Overload resolution: Don't calculate optional params into match score
(fixes #603)

This patch changes symbFindClosestOvlProc() to stop treating optional
parameters for which no args were provided as "full matches". I think it
makes more sense to score overloads only based on the given args, and
default args inserted for optional parameters shouldn't artifically push
an overload's score up.

It's an attempt to replace the "normalization" code added in bae05570
with a better solution, because the "matches /= total_args" division
is prone to rounding errors which can falsify the matching scores.

For example:
	declare sub f overload( ul as ulong, i as long )
	declare sub f overload( ull as ulongint, i as long )
	dim as byte b
	f( b, 1 )
The 1st overload scores matches=41, the 2nd matches=39. After the divison,
both have matches=20 though due to rounding, making them appear equal even
though the 1st one was originally better than the 2nd one.

Another example where it happens to work ok:
	declare sub f overload( ul as ulong, i as integer )
	declare sub f overload( ull as ulongint, i as integer )
	dim as byte b
	f( b, 1 )
The 1st overload scores matches=68, the 2nd matches=66. After the divison,
the 1st has matches=34 and the 2nd matches=33. Here the 1st one still is
better than the 2nd, so it doesn't make a difference.

Besides, it seems to me like the normalization code was only added to
work-around hCalcTypesDiff()'s problems with z/wstring params/args, judging
from the regression test added in bae05570. That problem should be fixed
now since the previous commit (2f9ea994) though, and now it should be
possible to remove the normalization code.

To change the behaviour, this patch
- removes the "matches += FB_OVLPROC_FULLMATCH" line from the optional
  parameter code path
- introduces a new local "is_match" variable, to be relied on instead of
  checking "matches = 0", because now that optional parameters no longer
  increase the "matches", a zero value in "matches" can be a valid match
  (in case of zero args given, matching an overload which has only optional
  parameters)