Overload resolution: Don't prefer paramless overload if zero args given
For example:

	declare f overload( )
	declare f overload( as integer = 0 )
	f( )  '' ambigious call

To be consistent, this should be an ambigious call, same as this:

	declare f overload( as integer )
	declare f overload( as integer, as integer = 0 )
	f( 0 )  '' ambigious call

C++ also behaves this way.

This change broke (or from my point of view: "exposed bugs in") some test
cases so I had to remove those. They had two default constructors, one
with zero parameters, one with an optional parameter - that should never
have been allowed, the compiler can't know which one to call. As in C++,
this is an "ambigious call" error now.