ast: Disable constant conversin overflow warnings during optimizations
Thanks to the changes from the 64bit branch, even optimizations use the
AST's constant folding to evaluate everything, and the AST was also changed
to check for overflows in constant operations more precisely. This lead
to it reporting warnings for code that looks ok, due to expressions being
changed by optimizations.

For example:
	dim as uinteger ui
	print ui - &hD800 - &hDC00
was "optimized" to:
	dim as uinteger ui
	print ui + 4294912000 + 4294910976
and then the constant accumulation optimizations caused the overflow
warning when trying to add those two huge values together.

It's worth noting that normal fbc (without the changes from 64bit branch)
doesn't show a warning when compiling either of the two snippets above,
while the new fbc from the 64bit branch would show a warning in both cases.

This patch adds a boolean to the AST context, allowing astOptimizeTree()
to temporarily disable the overflow warnings, so it can use astNewBOP()
etc. without worrying. This matches the idea of only showing warnings for
the user's code, but not for any internal expression restructurings etc.