Fix #663: Always treat base.method() as non-virtual call
BASE is a special keyword to explicitly access the base class and its
members. Assuming the base has a virtual method f() and the derived class
overrides it, doing BASE.f() from inside the derived class'es f() should
call the base's f(), instead of doing a virtual call (vtable lookup) and
calling the derived class'es f() recursively.

To implement this, it's necessary to add a boolean option to cProcArgList()
and co to allow callers to tell it whether to add a vtable lookup or not.
This boolean can be set to TRUE for parser function calls made from the
hBaseMemberAccess() functions. The boolean has to be propagated through
various middle-man functions like cUdtMember() and cMethodCall() but will
ultimately reach cProcArgList(). This patch adds a new parser flag called
FB_PARSEROPT_EXPLICITBASE for this purpose.

This way this option only applies to one statement or expression preceded
with "BASE." but not any other expressions in the same line as would
happen when adding the option to the global parser state.