Include namespaces into UDT name mangling (for storing into RTTI tables)
The Is operator is based around comparing the identifiers stored in RTTI
tables. GCC stores the mangled name, i.e. Foo is stored as "3Foo" and
ns::Foo is stored as "N2ns3FooE".

fbc didn't take into account namespaces in this scenario and always stored
"3Foo" no matter whether that UDT was in a namespace or not. This of course
prevented the Is operator from telling UDTs with the same identifier but
declared in different namespaces apart.