Add rtti/vtables to their parent UDT namespace, to fix name mangling
hBuildRtti() and hBuildVtable() used to do the mangling for rtti and vtable
manually and use the result as ALIAS for the global vars, but the problem
with that is that global vars inside namespaces, even if an ALIAS was
given, will additionally be mangled according to those namespaces, i.e.
the ALIAS is not the only thing respected; instead, more is added.

This of course caused rtti/vtables inside namespaces to have
GCC-incompatible name mangling. This patch fixes that by adding the
rtti/vtables to the parent UDT (like STATIC member variables) and marking
them with a new flag each, allowing hMangleVariable() to produce the
proper mangling.